Calculation2.javaファイルの作成と今までの演習内容をsrc直下へ
