+++ src/basicSyntax/Calculation2.java
@@ -1,0 +1,19 @@
+package basicSyntax;
+
+public class Calculation2 {
+	public static int addUp(int num1, int num2) {
+		return num1 + num2;
+	}
+
+	public static int subtract(int num1, int num2) {
+		return num1 - num2;
+	}
+	
+	public static int multiply(int num1, int num2) {
+		return num1 * num2;
+	}
+	
+	public static int divide(int num1, int num2) {
+		return num1 / num2;
+	}
+}
